extract some logic to model_processor
